requirements.txt
langchain
langchain-openai
langchain-community
faiss-cpu
python-dotenv
streamlit
pypdf
unstructured
sentence-transformers
scikit-learn
pytest
arxiv
beautifulsoup4
requests

